requirements.txt
Flask
docker-compose
gunicorn